2. Тесты для задачи "идеальный вес" [#159458 #164832]

+++ src/ru/job4j/calculator/Fit.java
@@ -1,16 +1,16 @@
 package ru.job4j.calculator;
 
 public class Fit {
-    public static double manWeight(float height) {
+    public static double manWeight(double height) {
         return (height - 100) * 1.15;
     }
 
-    public static double womanWeight(float height) {
+    public static double womanWeight(double height) {
         return (height - 110) * 1.15;
     }
 
     public static void main(String[] args) {
-        float height = 180;
+        double height = 180;
         double man = Fit.manWeight(height);
         System.out.println("Man 180 is " + man);
         double woman = Fit.womanWeight(height);
